frontend(cleaning): created delivery routing module and sorted imports
