+++ libs/material/src/lib/delivery/delivery.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RouterModule, Routes } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { TextFieldModule } from '@angular/cdk/text-field';
@@ -28,6 +27,7 @@ import { MovieMaterialListComponent } from './components/movie-material-list/mov
 import { MovieMaterialFormComponent } from './components/movie-material-form/movie-material-form.component';
 import { DeliveryInformationsStakeholdersComponent } from './components/delivery-informations-stakeholders/delivery-informations-stakeholders.component';
 
+import { DeliveryStatusesComponent } from './components/delivery-statuses/delivery-statuses.component';
 
 // Pages
 import { DeliveryListComponent } from './pages/delivery-list/delivery-list.component';
@@ -39,6 +39,8 @@ import { DeliveryAddChooseStarterComponent } from './pages/delivery-add-choose-s
 import { DeliveryAddTemplatePickerComponent } from './pages/delivery-add-template-picker/delivery-add-template-picker.component';
 import { DeliveryAddSettingsComponent } from './pages/delivery-add-settings/delivery-add-settings.component';
 import { DeliveryInformationsEditableComponent } from './pages/delivery-informations-editable/delivery-informations-editable.component';
+import { DeliveryAddSpecificDeliveryListPickerComponent } from './pages/delivery-add-specific-delivery-list-picker/delivery-add-specific-delivery-list-picker.component';
+import { DeliveryAddCompleteComponent } from './pages/delivery-add-complete/delivery-add-complete.component';
 import { MovieEditableComponent } from './pages/movie-editable/movie-editable.component';
 
 // Modules
@@ -56,6 +58,8 @@ import {
 } from '@blockframes/ui';
 import { OrganizationModule } from '@blockframes/organization';
 import { MaterialModule } from '../material/material.module';
+import { DeliveryRoutingModule } from './delivery-routing-module';
+import { MovieModule } from '@blockframes/movie';
 
 // Material
 import { MatCardModule } from '@angular/material/card';
@@ -79,142 +83,6 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatSortModule } from '@angular/material/sort';
 import { MatStepperModule } from '@angular/material/stepper';
-
-// Guards
-import { MovieActiveGuard, MovieListGuard, MovieModule } from '@blockframes/movie';
-import { DeliveryActiveGuard } from './guards/delivery-active.guard';
-import { DeliveryListGuard } from './guards/delivery-list.guard';
-import { DeliveryMaterialsGuard, MovieMaterialsGuard } from '../material';
-import { TemplateActiveGuard } from '../template/guards/template-active.guard';
-import { TemplateListGuard } from '../template/guards/template-list.guard';
-import { DeliveryAddSpecificDeliveryListPickerComponent } from './pages/delivery-add-specific-delivery-list-picker/delivery-add-specific-delivery-list-picker.component';
-import { DeliveryAddCompleteComponent } from './pages/delivery-add-complete/delivery-add-complete.component';
-import { DeliveryStatusesComponent } from './components/delivery-statuses/delivery-statuses.component';
-
-const routes: Routes = [
-  {
-    path: 'add',
-    canActivate: [MovieListGuard],
-    canDeactivate: [MovieListGuard],
-    children: [
-      {
-        path: '',
-        pathMatch: 'full',
-        redirectTo: '1-find-movie'
-      },
-      {
-        path: '1-find-movie',
-        pathMatch: 'full',
-        component: DeliveryAddFindMovieComponent
-      },
-      {
-        path: ':movieId',
-        canActivate: [MovieActiveGuard],
-        canDeactivate: [MovieActiveGuard],
-        children: [
-          {
-            path: '2-choose-starter',
-            pathMatch: 'full',
-            component: DeliveryAddChooseStarterComponent
-          },
-          {
-            path: '3-pick-template',
-            canActivate: [TemplateListGuard],
-            canDeactivate: [TemplateListGuard],
-            pathMatch: 'full',
-            component: DeliveryAddTemplatePickerComponent
-          },
-          {
-            path: '3-pick-specific-delivery-list',
-            pathMatch: 'full',
-            component: DeliveryAddSpecificDeliveryListPickerComponent
-          },
-          {
-            path: '4-settings',
-            pathMatch: 'full',
-            component: DeliveryAddSettingsComponent
-          },
-          {
-            path: '5-complete',
-            pathMatch: 'full',
-            component: DeliveryAddCompleteComponent
-          },
-          {
-            path: ':templateId',
-            canActivate: [TemplateActiveGuard],
-            canDeactivate: [TemplateActiveGuard],
-            children: [
-              {
-                path: '4-settings',
-                pathMatch: 'full',
-                component: DeliveryAddSettingsComponent
-              }
-            ]
-          }
-        ]
-      }
-    ]
-  },
-  {
-    path: ':movieId',
-    canActivate: [MovieActiveGuard],
-    canDeactivate: [MovieActiveGuard],
-    children: [
-      {
-        path: '',
-        redirectTo: 'list',
-        pathMatch: 'full'
-      },
-      {
-        path: 'list',
-        canActivate: [DeliveryListGuard],
-        canDeactivate: [DeliveryListGuard],
-        component: DeliveryListComponent
-      },
-      {
-        path: 'template-picker',
-        // TODO: Getting redirected to templates/list if there is no template to load => ISSUE#648
-        canActivate: [TemplateListGuard],
-        canDeactivate: [TemplateListGuard],
-        component: DeliveryTemplateListComponent
-      },
-      {
-        path: 'materials',
-        canActivate: [MovieMaterialsGuard],
-        canDeactivate: [MovieMaterialsGuard],
-        component: MovieEditableComponent
-      },
-      {
-        path: ':deliveryId',
-        canActivate: [DeliveryActiveGuard],
-        canDeactivate: [DeliveryActiveGuard],
-        children: [
-          {
-            path: '',
-            redirectTo: 'view',
-            pathMatch: 'full'
-          },
-          {
-            path: 'list',
-            canActivate: [DeliveryMaterialsGuard],
-            canDeactivate: [DeliveryMaterialsGuard],
-            component: DeliveryEditableComponent
-          },
-          {
-            path: 'stakeholders',
-            component: DeliveryTeamworkEditableComponent
-          },
-          {
-            path: 'informations',
-            canActivate: [DeliveryMaterialsGuard],
-            canDeactivate: [DeliveryMaterialsGuard],
-            component: DeliveryInformationsEditableComponent
-          }
-        ]
-      }
-    ]
-  }
-];
 
 @NgModule({
   declarations: [
@@ -270,6 +138,9 @@ const routes: Routes = [
     TeamWorkModule,
     OrganizationModule,
     EditableSidenavModule,
+    DeliveryRoutingModule,
+    MovieModule,
+    FeedbackMessageModule,
 
     // Material
     MatExpansionModule,
@@ -293,10 +164,6 @@ const routes: Routes = [
     MatTooltipModule,
     MatAutocompleteModule,
     MatSortModule,
-
-    RouterModule.forChild(routes),
-    MovieModule,
-    FeedbackMessageModule,
     MatStepperModule
   ],
   entryComponents: [NewTemplateComponent, DeliverySignComponent]
